admin-pager/requirements.txt
sendgrid
python-decouple
